requirements.txt
pytest
numpy
pandas
apache-airflow-providers-google
google-cloud-storage
protobuf
shap
mlflow
fairlearn
scipy
xgboost
seaborn
matplotlib
kneed
scikit-learn